+++ src/Controler/index.js
@@ -3,19 +3,22 @@ import { components } from "../View/index.js"
 const changeView = (route) => {
     const container = document.querySelector("#app");
     container.innerHTML = "";
-    switch(route) {
+    switch (route) {
         case "":
-                { return container.appendChild(components.welcome()) }
-        case "#/": 
-        { return container.appendChild(components.welcome()) }
-        case "#/register": 
-        {return container.appendChild(components.register()) }
+            { return container.appendChild(components.welcome()) }
+            break
+        case "#/":
+            { return container.appendChild(components.welcome()) }
+            break
+        case "#/register":
+            { return container.appendChild(components.register()) }
+            break
         case "#/home":
-        { return container.appendChild(components.home()) }
-        default:
-            break;
+            { return container.appendChild(components.home()) }
+            break
     }
     console.log(route)
 };
+
 
 export { changeView }
