+++ src/Model/firebase.js
@@ -1,6 +1,6 @@
 /*-------------------------FIREBASE KEY AND CONFIG-------------------------*/
 
-var firebaseConfig = {
+let firebaseConfig = {
     apiKey: "AIzaSyDZBy9n50HCcJmEOL5-zzYyguPmUJGk3yM",
     authDomain: "red-feminista.firebaseapp.com",
     databaseURL: "https://red-feminista.firebaseio.com",
@@ -12,18 +12,18 @@ var firebaseConfig = {
 };
 // Initialize Firebase
 firebase.initializeApp(firebaseConfig);
-//const dataBase = firebase.firestore();
+let dataBase = firebase.database();
 
 /*------------------------------LOG IN -----------------------------*/
 
-function logInFn(email, password){
+function logInFn(email, password) {
     firebase.auth().signInWithEmailAndPassword(email, password).catch(function(error) {
         // Handle Errors here.
-        var errorCode = error.code;
-        var errorMessage = error.message;
+        let errorCode = error.code;
+        let errorMessage = error.message;
         // ...
-        alert("ERROR: " + errorMessage );
-      }); 
+        alert("ERROR: " + errorMessage);
+    });
 };
 
 //This is for view changes. We are not using it yet. 
@@ -48,31 +48,32 @@ function logInFn(email, password){
 
 /*------------------------------SIGN UP -----------------------------*/
 
-function signUpFn (email, password) {
+function signUpFn(email, password) {
     firebase.auth().createUserWithEmailAndPassword(email, password).catch(function(error) {
         var errorCode = error.code;
         var errorMessage = error.message;
         console.log(errorCode);
         console.log(errorMessage);
-        alert("Tu contraseña o tu correo no es válido" +" "+ ":(");
+        alert("Tu contraseña o tu correo no es válido" + " " + ":(");
     });
 };
 
 /*------------------------------LOG OUT-----------------------------*/
 
-function logOutFn(){
-    firebase.auth().signOut();/*.then(function() {
-        // Sign-out successful.
-        document.querySelector("#createAccount").style.display = "block";
-        document.querySelector("#welcomeUser").style.display = "none";
-      }).catch(function(error) {
-        // An error happened.
-    });*/
-}; 
+function logOutFn() {
+    firebase.auth().signOut();
+    /*.then(function() {
+            // Sign-out successful.
+            document.querySelector("#createAccount").style.display = "block";
+            document.querySelector("#welcomeUser").style.display = "none";
+          }).catch(function(error) {
+            // An error happened.
+        });*/
+};
 
 
 /*------------------------------FACEBOOK AUTH-----------------------------*/
-function facebook(){
+function facebook() {
     var provider = new firebase.auth.FacebookAuthProvider();
     provider.addScope('user_birthday');
     firebase.auth().useDeviceLanguage();
@@ -82,8 +83,8 @@ function facebook(){
         // The signed-in user info.
         var user = result.user;
         console.log(user)
-        // ...
-      }).catch(function(error) {
+            // ...
+    }).catch(function(error) {
         // Handle Errors here.
         var errorCode = error.code;
         var errorMessage = error.message;
@@ -92,7 +93,7 @@ function facebook(){
         // The firebase.auth.AuthCredential type that was used.
         var credential = error.credential;
         // ...
-      });
+    });
 };
 
 /*------------------------------GOOGLE AUTH-----------------------------*/
@@ -123,25 +124,25 @@ function ponUp() {
 /*------------------------------TWITTER AUTH-----------------------------*/
 
 
-function twitter(){
-  var provider = new firebase.auth.TwitterAuthProvider();
-  firebase.auth().useDeviceLanguage();
-  firebase.auth().signInWithPopup(provider).then(function(result) {
-      // This gives you a Twitter Access Token. You can use it to access the Twitter API.
-      var token = result.credential.accessToken;
-      // The signed-in user info.
-      var user = result.user;
-      console.log(user)
-      // ...
+function twitter() {
+    var provider = new firebase.auth.TwitterAuthProvider();
+    firebase.auth().useDeviceLanguage();
+    firebase.auth().signInWithPopup(provider).then(function(result) {
+        // This gives you a Twitter Access Token. You can use it to access the Twitter API.
+        var token = result.credential.accessToken;
+        // The signed-in user info.
+        var user = result.user;
+        console.log(user)
+            // ...
     }).catch(function(error) {
-      // Handle Errors here.
-      var errorCode = error.code;
-      var errorMessage = error.message;
-      // The email of the user's account used.
-      var email = error.email;
-      // The firebase.auth.AuthCredential type that was used.
-      var credential = error.credential;
-      // ...
+        // Handle Errors here.
+        var errorCode = error.code;
+        var errorMessage = error.message;
+        // The email of the user's account used.
+        var email = error.email;
+        // The firebase.auth.AuthCredential type that was used.
+        var credential = error.credential;
+        // ...
     });
 };
