+++ src/View/index.js
@@ -1,11 +1,13 @@
 import Welcome from "./welcome.js";
 import Register from "./register.js";
 import Home from "./home.js";
+import Perfil from "./perfil.js"
 
 const components = {
     welcome: Welcome,
     register: Register,
-    home: Home
+    home: Home,
+    perfil: Perfil
 };
 
 export { components };
