+++ src/View/perfil.js
@@ -1,0 +1,10 @@
+export default () => {
+    const viewHome = `
+        
+    `
+
+    const divElmnt = document.createElement("div")
+    divElmnt.innerHTML = viewHome;
+
+    return divElmnt;
+};
