--- src/firebase.js
@@ -1,14 +1,0 @@
-var firebaseConfig = {
-    apiKey: "AIzaSyDZBy9n50HCcJmEOL5-zzYyguPmUJGk3yM",
-    authDomain: "red-feminista.firebaseapp.com",
-    databaseURL: "https://red-feminista.firebaseio.com",
-    projectId: "red-feminista",
-    storageBucket: "red-feminista.appspot.com",
-    messagingSenderId: "7507219044",
-    appId: "1:7507219044:web:bc69a1f4bd8ea5a1218571",
-    measurementId: "G-2L40V7GZPX"
-};
-firebase.initializeApp(firebaseConfig);
-//const dataBase = firebase.firestore();
-let database = firebase.database();
-//firebase.analytics();
